pyproject.toml
[project]
name = "knowledge-flow-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "crewai-tools>=1.5.0",
    "google-adk>=1.19.0",
    "googlesearch-python>=1.3.0",
    "python-dotenv>=1.2.1",
]
